Add R flower art project
R is a tool for doing serious statistics and data analysis. But not everything in life can be serious, life is also beautiful, and R can make beautiful things too. R can make art. The arrangement of leaves on a plant stem is ruled by spirals. This fact is called phyllotaxis and it is a nice example of how mathematics can describe patterns in nature. In this project, I will invent flowers using this fact.